fcc-aas-alpha-proto/webapp/src/app/layouts/navbar/navbar.component.ts
import { Component } from '@angular/core';
import { AuthService } from 'src/app/core/auth/auth.service';

@Component({
  selector: 'aas-navbar',
  templateUrl: './navbar.component.html',
  styleUrls: ['./navbar.component.scss']
})
export class NavbarComponent {
  allowNavigation = true;

  constructor( private authService: AuthService){}

  isAuthenticated(): boolean {
    return this.authService.isAuthenticated() as boolean;
  }

  signIn(): void {
    this.authService.signIn();
  }

  signOut(): void {
    this.authService.signOut();
  }
}
